Creating public API for robot info, working through uses in manager
